refactor: use package for magefiles
