+++ magefiles/magefile.go
@@ -1,4 +1,8 @@
 //go:build mage
+// +build mage
+
+// This file is maintained by the plugin sdk tooling.
+// Please do not make changes to this file.
 
 package main
 
@@ -6,18 +10,6 @@ import (
 	//mage:import
 	"github.com/mattermost/mattermost-plugin-legal-hold/plugin_magefile"
 )
-
-// RegisterAdditionalBinaries allows plugins to register additional binaries to be built
-func init() {
-	// Register the processor binary
-	plugin_magefile.RegisterBinary(plugin_magefile.BinaryBuildConfig{
-		Name:             "processor",
-		OutputPath:       "./bin",
-		BinaryNameFormat: "processor-{{.Manifest.Version}}-{{.GOOS}}-{{.GOARCH}}",
-		WorkingDir:       "./processor",
-		PackagePath:      ".",
-	})
-}
 
 // Aliases defines some targets migrated from the old Makefile to ease the transition to mage
 // You can find the default aliases in plugin_magefile/init.go
